refactor: adicionei o método save no computerRepository
