+++ Repositories/ComputerRepository.cs
@@ -33,4 +33,19 @@ class ComputerRepository
         return computers;
     }
 
+    public void Save(Computer computer)
+    {
+        var connection =  new SqliteConnection("Data Source=database_test.db");
+        connection.Open();
+
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);";
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+        
+        command.ExecuteNonQuery();
+        connection.Close();
+
+    }
 }
